admin/pyproject.toml
[tool.poetry]
name = "StranaAdmin"
version = "0.1.0"
description = "Администрирование личного кабинета Страна Деволпмент"
authors = ["Bogdan <bb@idaporject.com>"]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.8.5"
boto3 = "~1.13.20"
Pillow-SIMD = "~7.0.0.post3"
gunicorn = "~20.0.4"
requests = "~2.23.0"
phonenumberslite = "~8.12.4"
django-phonenumber-field ="~4.0.0"
django-hashedfilenamestorage = {git = "https://github.com/ecometrica/django-hashedfilenamestorage.git", rev = "9a3b1d1b74ae96a84dac4cf5bd1c9b6dfce2d7a3"} # При установки из pipy неверно вычисляются зависимости
django-storages = "~1.9.1"
django-ckeditor = "~5.9.0"
django-imagekit = {git = "https://github.com/matthewwithanm/django-imagekit.git", rev = "3317273401d65b1d2c70c635e22132998f957290"} # Совместимость с Django 2.2
django-ajax-image-upload = "~0.9.4"
django-admin-sortable2 = "~0.7.6"
django-admin-rangefilter = "~0.6.0"
django-admin-list-filter-dropdown = "~1.0.3"
django-select2 = "~7.4.2"
django-multi-email-field = "~0.6.1"
django-colorful = "~1.3.0"
django-nested-admin = "^3.3.3"
django-redis-cache = "^3.0.0"
django = "3.1"
sentry-sdk = "^1.0.0"
django-grappelli = "^3.0.5"
django-admincharts = "^0.4.1"
python-dotenv = "^1.0.0"
django-solo = "^1.2.0"
django-debug-toolbar = "3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
